Create Using recursion
We build a function using recursive method to calculate the value of a number in the Fibonacci Sequence. However, this method becomes very slow when calculating a range of numbers greater than 50 as each time the function tries to process a new number it has to go through this For-loop which start from 0, therefore calculating with this method uses a lot of memory and can do harm to your computer.