Create Using Iterables
This method uses iterables  instead of recursion. 
[0, 1, 1, 2, 3, 5, 8]
This is a Fibonacci Sequence in range(7). We can see that in every three numbers , the third number is the sum of the first two; and after each calculation, the original first number becomes the original second number, and the original second number has now become the original third number. So we only need to change the iterable objects to achieve this function.